requirements.txt
pymongo
python-dotenv
sentence-transformers